Function sonar_desktop_app::get_interfaces::get_interfaces

pub fn get_interfaces() -> Vec<String>

Récupère les noms de toutes les interfaces réseau sur le système, avec une entrée supplémentaire pour représenter la sélection de toutes les interfaces.

Cette fonction se sert de pnet::datalink::interfaces pour obtenir une liste de toutes les interfaces réseau disponibles sur le système courant. Pour Linux, elle retourne les noms d’interface et pour Windows, elle retourne les adresses MAC des interfaces. Enfin, elle ajoute une chaîne “Toutes les interfaces” à ce vecteur, permettant de représenter l’option de choisir toutes les interfaces dans une interface utilisateur ou un paramètre de configuration.

Retours

Un Vec<String> contenant les noms ou adresses MAC de toutes les interfaces réseau trouvées sur le système, plus une entrée “Toutes les interfaces” indiquant l’option de sélection de toutes les interfaces.

Exemples

Utilisation simple :

use sonar_desktop_app::get_interfaces::get_interfaces;

let interface_names = get_interfaces();
for name in interface_names {
    println!("{}", name);
}