Struct sonar_desktop_app::tauri_state::SonarState

pub struct SonarState {
    pub matrice: HashMap<PacketInfos, u32>,
    pub filter_ipv6: bool,
    pub actif: bool,
}

SonarState encapsule l’état global de l’application Sonar.

Cette structure est conçue pour stocker et gérer les informations sur les trames réseau capturées, y compris le comptage de leurs occurrences.

Structure

SonarState contient un Arc<Mutex<HashMap<PacketInfos, u32>>>.

Arc permet un accès thread-safe et partagé à l’état.

Mutex garantit que l’accès à l’état est mutuellement exclusif, empêchant les conditions de concurrence.

HashMap<PacketInfos, u32> stocke les trames réseau (PacketInfos) et leur nombre d’occurrences (u32).

Exemple

use std::sync::{Mutex, Arc};
use std::collections::HashMap;
use crate::sniff::capture_packet::layer_2_infos::PacketInfos;
use crate::SonarState;

let state = SonarState::new();
// Utilisez `state` ici pour gérer les trames réseau et leur comptage

Fields
matrice: HashMap<PacketInfos, u32>filter_ipv6: boolactif: bool
Implementations

impl SonarState

pub fn new() -> Arc<Mutex<Self>>

pub fn toggle_actif(&self)

pub fn toggle_filter_ipv6(&self)

pub fn get_matrice(&self) -> &HashMap<PacketInfos, u32>

pub fn update_matrice_with_packet(&mut self, new_packet: PacketInfos)

pub fn cmd_save_packets_to_csv(&self, file_path: String) -> Result<(), MyError>

Fonction pour enregistrer les paquets vers un fichier CSV.

Arguments

file_path - Chemin du fichier CSV.

app - Application handle de Tauri.

Exemple

cmd_save_packets_to_csv(String::from("paquets.csv"), state);

pub fn cmd_save_packets_to_excel( &self, file_path: String, ) -> Result<(), MyError>

Fonction pour enregistrer les paquets vers un fichier Excel.

Arguments

file_path - Chemin du fichier Excel.

state - État contenant les données des paquets.

Exemple

cmd_save_packets_to_excel(String::from("paquets.xlsx"), state);

pub fn get_matrice_data(&self) -> Result<String, String>

Récupère et sérialise les données de trafic réseau depuis l’état partagé.

Retour

Cette fonction retourne Ok(String) contenant les données sérialisées en cas de succès, ou Err(String) avec un message d’erreur en cas d’échec.

Exemples

let result = get_matrice_data(app);
match result {
    Ok(json_string) => println!("Données sérialisées : {}", json_string),
    Err(e) => eprintln!("Erreur : {}", e),
}

pub fn get_graph_data(&self) -> Result<String, String>